+++ CHANGELOG.md
@@ -2,8 +2,6 @@
 
 All notable changes to the "materialdesignicons-intellisense" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
 ## 0.0.1
 
 Initial release of Material Design Icons Intellisense
@@ -126,3 +124,15 @@ Initial release of Material Design Icons Intellisense
 ## 2.0.2
 
 - Improved readme
+
+## 3.0.0
+
+- Removed settings
+  - `materialdesigniconsIntellisense.insertPrefix`
+  - `materialdesigniconsIntellisense.insertSuffix`
+- Changed settings
+  - `materialdesigniconsIntellisense.insertStyle`
+- Added settings
+  - `materialdesigniconsIntellisense.matchers`
+- Added commands
+  - `materialdesigniconsIntellisense.changeSettings`